new flag -f to enable a faulthandler log
Enabling this will create a "fault_logs" directory with a new log file
every time you start the application with this flag. If an exception
then causes the application to crash, a traceback of all threads will
be dumped there. For some reason, faulty and dying hardware has shown
itself to me as random crashes of Happypanda with the most bizarre
exceptions and from lines of code where those exceptions make
absolutely no sense. Exceptions like "memory access violation" in a
line where the app is creating an empty list or "too many arguments"
one time and "not enough arguments" another time for the same line of
code that calls a parameterless method.