gallery artists "Anthology" or "アンソロジー" will not be auto-replaced
If the title parser returns "Anthology" as the artist, then neither the
metadata file nor the E-Hentai tag fetch will replace it unless it's
with another variant of "Anthology". Replacing the gallery artist with
the first one found in the tags can make sense but not for anthologies
and I hope I found every spot where this can happen.